CLIENT-4360 Create index to use “integer” instead of “numeric” starting with Aerospike server >= 8.1.2

+++ client/src/test/java/com/aerospike/client/fluent/query/ExpSecondaryIndexTest.java
@@ -94,7 +94,7 @@ public class ExpSecondaryIndexTest extends ClusterTest {
 
 	public static void addExpSI() {
 		try {
-			session.createIndex(dataSet, indexName, IndexType.NUMERIC, IndexCollectionType.DEFAULT, exp)
+			session.createIndex(dataSet, indexName, IndexType.INTEGER, IndexCollectionType.DEFAULT, exp)
 				.waitTillComplete();
 		}
 		catch (AerospikeException ae) {
