+++ client/src/test/java/com/aerospike/client/fluent/query/QueryFilterExpTest.java
@@ -56,7 +56,7 @@ public class QueryFilterExpTest extends ClusterTest {
 		}
 		
 		try {
-			session.createIndex(dataSet, indexName, binName, IndexType.NUMERIC, IndexCollectionType.DEFAULT)
+			session.createIndex(dataSet, indexName, binName, IndexType.INTEGER, IndexCollectionType.DEFAULT)
 				.waitTillComplete();
 		} catch (AerospikeException ae) {
 			if (ae.getResultCode() != ResultCode.INDEX_ALREADY_EXISTS) {
